requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.2
pydantic[email]
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
plotly==5.18.0
pandas==2.1.3
email-validator==2.1.0.post1 